[Enhancement] Compaction scheduling based on compaction score(1/3)
This PR is part of the PR stack to support compaction scheduling based
on the compaction score of each tablet.

Instead of scheduling compactions based on the number of versions, the
new compaction policy will collect the compaction score of each tablet
and make decisions based on the average and median compaction score.