[Enhancement] Compaction scheduling based on compaction score(2/3)
This PR is part of the PR stack to support compaction scheduling
based on the compaction score of each tablet.

Instead of scheduling compactions based on the number of versions, the
new compaction policy will collect the compaction score of each tablet
and make decisions based on the average and median compaction score:
compact a partition if the average compaction score or the median
compaction score of the partition is greater than or equals to
`lake_compaction_score_selector_min_score`(configurable).