[BugFix] BE crash on illegal JSON (#12850)
When the input is invalid JSON, like `{"key": 1} "key": 2}`, the 2nd document would not be parsed successfully. In order to get the context that leads the parsing failure, we use `JsonFunctions::to_json_string(doc)` to print the content of the doc. But for doc that is not parsed successfully, the  `JsonFunctions::to_json_string(doc)` would lead the BE crash.
This PR removes the printing of not parsed doc. The BE would record the current parsed location, and print the not-parsed content to the error log. In this way, not only the bug leading the crash would be resolved, but also the user is easier to get to know the error in the json input.
Besides, this PR would not change the behavior of the SR. It would modify the error message and adds more content to the error log file.