[BugFix] Fix incorrect dla query results for parquet with case-sensitive predicates. (#13698)
For lake source which is case sensitive, the column names may be upper-case.
When querying the parquet file only with an upper predicate, as we didn't deal with
the predicate cases completely, it will be treated as `not_existed_slots` and the
query will be skipped, which causes an empty result.

(cherry-pick from 70be355269ec619d8f6742cfd)