[Refactor] Remove Field dependency in ZoneMap (#13870)
Before this CL, ZoneMap uses Field to visit a raw memory as the min/max state for the writing page and segment.

To speed-up writing performance for CHAR/VARCHAR, Zone Map writer will allocate memory before, then when the value is updated, it can copy the content directly without allocating new memory content. So it needs the CHAR/VARCHAR length information in its corresponding Field.

Because we want to delete the Field struct, I introduce the ZoneMapDatum struct in this PR. This struct is a self-contained struct, for string type, it will allocate memory through std::string member rather than MemPool.

Because ZoneMap is already a template class, so I implement ZoneMapDatum in a template way too, which will have better performance and easier accessing way.

The original unit test pass and I tested it with some designed case, both works fine.