[Enhancement] Add BE down tolerate time for colcoate balance (#13925)
Colocate balance is very time-and-resource-consuming, but when BE is down for a little time, colcoate balance will be triggered. We should avoid colocate balance as much as possible. So add a config param tablet_sched_colocate_be_down_tolerate_time_s, if BE is down beyond this time, tablets on that BE of colcoate table will be migrated to other available BEs.

(cherry picked from commit 803ae6eaf4537645d9accdb82d74f4beceff0dc0)