Reduce the memory usage of small segment file iterator (#13861)
If a BE hangs up for some time, it will copy incremental versions from other BEs and execute compaction. The current primary key model does not limit the number of compaction versions, and SegmentIterator will reserve unnecessary memory. If there are many versions, each segment is small, and there are many columns, a large amount of memory will be consumed. In the user's scenario, 10,000 versions and one wide table consume 100G+ memory.

In my test, the pr can reduce the memory usage from 3.2G to 200M in the test case described in issue #13873

The current repair method is not elegant, and the allocation/release logic of `Column` on the storage layer is quite chaotic, so it will be optimized after careful consideration.

In my performance test, there is no impact on the compaction performance of large files or small files

Q: Why not modify the `_chunk_size `on `ChunkIterator` ?
A: Because the `_chunk_size`, will be used by upper-level `ChunkIterator` such as `AggregateIterar`.

Q: if `enable_segment_overflow_read_chunk` why reserve more memory for a small segment file?
A: In fact, it is not necessary, mainly because of code unification.