[Enhancement] Split tablets into small batches to decrease db lock occupation (#13070) (#14033)
TabletChecker/ColocateTableBalancer/DiskAndTabletLoadReBalancer scan all tablets to repair or balance tablet, and the db lock is occupied all the check time. If the number of tablet is large, lock conflicts can be frequent, and the writeLock request will have to wait for long time.
Divide all partitions into small batch to check tablets, so that db.readLock can be released between batches.