[BugFix] bugfix_get_empty_tableNames_in_dropDb (#13797)
Problem:
This Problem was introduced by commit bdb4da1. This commit use Collections.unmodifiableSet instead of new HashSet to reduce the memory usage of fe. But the problem here is that Collections.unmodifiableSet is just return the view of nameToTable.keySet(). Before use the TableNames in recycleDatabase, nameToTable had been dropped and also the view of it would become empty, which cause this problem.

Solution:
We copy the nameToTable.keySet() only when drop the database.