[Enhancement] Optimize thrift server pool  (#14571)
The current implementation of thrift server is blocking IO, that means a connection will occupy one thread, if the request number is more that the worker thread, some requests will be rejected. In order to release the worker thread as soon as possible, we can set the connection idle time to a small value. Since the worker thread can be released for the idle connection, the server can pend the extra request to a queue.