[cherry-pick][branch-2.2][Feature] Support configure rocksdb meta column family options. (#14398)
The default rocksdb block cache size is 8M, which may be too small. It should be configured.
just like block cache size, other column family options should also be configured.
This PR adds a configuration in be.conf to deliver a string to rocksdb GetColumnFamilyOptionsFromString API to parse.
Besides, prefix_extractor and compression are fixed options in case of incorrect.