[Refactor] Make aggregate method return lowercase name (#15031)
This is a previous CL for later changes. In this CL I changed the get_string_by_aggregate_method() to return a lower-case function name, which can be used to find AggregateFunction in FunctionResolver.
I have checked in the earlier version all aggregate_method read from storage will be converted to upper. So this change doesn't influence the compatibility.