[BugFix] unstable case for backup(#15097) (#15098)
Problem:
exec the BackupPrimaryKeyTest and BackupTest concurrently, sometimes the metafile is not found.
because this two cases use the same file path to save the meta data. If one of them finished first
the directory may be deleted and the other on may lost its meta data.

Solution:
Change the directory of BackupPrimaryKeyTest for saving meta file.