[BugFix] Fix memory leak for insert load job
The createLoadJob function adds all the load job to CallbackFactory, and it will be removed when the job is finished(or cancelled). But the InsertLoadJob is a sync job, and only be recored to LoadManager when finished, so there is no state transfer for the InsertLoadJob, and cause it will never be removed from CallbackFactory.
To fix this bug, we should only add the unfinished job to CallbackFactory.