[BugFix] Update Statement fail when SET a column using DEFAULT key word(#15181) (#15182)
Problem:
When we Update a column which is defined the default value, if we use set = default in update statement, we will get
unknow error because the DefaultValueExpr is not handled properly.

Solution:
Add the some code to cast the DefaultValueExpr to corresponding Literal type.

(cherry picked from commit e3e2d7125e06cde71a2830a172957454c3525a7e)

# Conflicts:
#	fe/fe-core/src/main/java/com/starrocks/sql/analyzer/UpdateAnalyzer.java