[BugFix] s3 broker load(without broker) bug
If we deploy the starrocks on ECS(aliyun), and ingest data from S3 through broker load(without broker),
broker load will fail, error msg is 'empty response body'.
This is because we didn't set the region field for s3-client(c++ sdk), in this pr, we parse the region field
from the endpoint in FE, and send the region to BE, then BE will set the region for the s3-client.

Signed-off-by: xyz <a997647204@gmail.com>