[BugFix] Fix Potential out of bounds or heap overflow in encode procedure when primary key model compaction with sort key. (#15124)
the old code use the number of primary keys as encode columns' number, should not still use this way.
for example, when sort keys exist, primary keys indexes are {0}, the sort keys indexes are {1, 2}, use the size of primary keys indexes will be 1, in the function prepare_ops_datas, will encode the 1 prefix of sort keys indexes, it will be {1}, it is not right, should be {1, 2}. this PR change the function to use an indexes instead of the prefix N indexes.
and, added some new corresponding case to expose this bug.

./run-ut.sh --run --gtest_filter=TabletUpdatesTest.horizontal_compaction_with_sort_key

(cherry picked from commit 452376e489317243ae7cacce139d11cb3ac8be3a)

# Conflicts:
#	be/src/storage/primary_key_encoder.cpp
#	be/src/storage/primary_key_encoder.h
#	be/src/storage/rowset_merger.cpp