load support fast cancel (#15514)
If the load task times out or the user terminates early, the resources used by load cannot be released immediately.
So we need to support FastCancel and release resources as soon as possible.
This is the first pr: Modify the implement of cancel, only set status, will not be blocked by I/O, queue or other behavior.