[BugFix] Fix backend not found bug when backend is alive (#15811)
If backend status is changed to dead(alive=false), it will be added to CoordinatorMonitor, and all the running coordinators(load or query) related to that backend will be cancelled. But CoordinatorMonitor is a singleton object, checkpointer which replays history log and generates image will also update CoordinatorMonitor, it will cause the backend is marked to dead by mistake.
To fix this bug, we should reject the checkpoint thread to update CoordinatorMonitor.

(cherry picked from commit b0ab86838839aa6c25950ac558e7ce3e7963570c)