[Bugfix] Fix TAuthenticateParams incompatible with the old version (#16130)
In #11871, to support authentication for StarRocks external table, we add three fields in thrift struct TAuthenticateParams: host, db_name and table_names, and they are marked required which leads to a compatible problem. Denote the starrocks before #11871 as v1, and starrocks after #11871 as v2, and we can ingesting data to v1 via the starrocks external table on v2 , but can't ingest data to v2 via external table on v1. The exception is as follows

#11871 adds host, db_name and table_names to TAuthenticateParams as required

struct TAuthenticateParams {
    1: required string user
    2: required string passwd
    3: required string host
    4: required string db_name
    5: required list<string> table_names;
}
The reason is that

external table on v1 will issue a RPC to v2 to request the meta of destination table
The RPC carries a parameter TAuthenticateParams created on v1 which not contains fields host, db_name and table_names
thrift requires these fields on v2, so the RPC failed
when ingesting data from v2 to v1, the RPC created on v2 does not carry TAuthenticateParams, so there is no problem

(cherry picked from commit 99b2b77f871d4107d1ae60be572e327a3fff0463)