[Refactor] Use AggregateFunction in storage layer  (#15081)
Currently, there are two implementation in both storage layer and
computation layer of aggregate function, such as sum, min, max,
bitmap_union, hll_union and percentile_union. There are some bad
influences of this kind of implementation
1. similiar code are implemented in two different places
2. difficult to add new aggregate function. When new aggregate function
   is added, these two places must be changed too.
3. some already exists computation aggregate functions can't be used by
   storage layer, and can't be used in aggregate model. For example:
   groupby_concat

After this change, most of aggregate functions can be used in Aggregate
Key model as long as it has the same input and output type. Even UDAF
can be supported in the future.