[Enhancement] Optimize ExportExportingSubTask Retry Strategy
For olap table, it may have multiple replica.
Current ExportExportingSubTask retry will run the same coord(plan fragment) again, however, if the tablet we want
has already been reblanced, or if the tablet version we want has been compacted, retry will definitely fail because
we will still try to load the same tablet.
To increase the probability of success of retry, we need reset the coord, generate a new plan fragment,
and select a new replica for each tablet based on the latest distribution.

Signed-off-by: xyz <a997647204@gmail.com>