[BugFix] only set not null for primary key column (#16431)
For now, we cannot obatin the correct nullable info from both logical plan and excutable plan, so we cannot define the meta of cols in the create table statement based on the output cols of the query.
So we take a loose restriction for nullable info. Only cols in the primary key are non-nullable and the other cols are all nullable.

(cherry picked from commit 4a67fa06711e2bd168a4967defb0aac61dda28c0)

# Conflicts:
#	fe/fe-core/src/main/java/com/starrocks/sql/analyzer/CTASAnalyzer.java
#	fe/fe-core/src/test/java/com/starrocks/sql/analyzer/CTASAnalyzerTest.java