[BugFix] Fix the version fall back when the RPC lag (#16841)
If the replica version in the FE catches up to 10, it should reject the smaller version(less than 10) reported by RPC.
The PRC may lag due to the high traffic network jam.  If the report is not rejected, the version falling back results in the commit error like the following.
```
 Commit failed. txn: 1002 table: test1 tablet: 10016 quorum: 0<1 errorReplicas: 10017:{be:10002 172.26.92.139 V:9 LFV:10},]}}, msg={replay end load job}
```