[Refactor] Wait for apply finish when exec primary key publish version (#17101)
PrimaryKey table only submit apply task to thread pool and return success directly when we do publish version right now. If we query immediately after publish success, we may need to wait for version apply finish which cause unstable query time.
So we try to wait for apply finished when we exec publish version.

It is important to note that this solution is a temporary one and does not completely solve the problem of unstable query time, i will reconsider the primary key publish version mechanism and solve the problem in the future pr.

Signed-off-by: zhangqiang <qiangzh95@gmail.com>
Signed-off-by: zhangqiang <qiangzh95@gmail.com>