[BugFix] Fix the problem of slow `show frontends` when the Kubernetes pod restart (#17341)
Fixes #17339
The bdb rpc client ReplicationGroupAdmin converts the FQDN to ip after creation, but the ip of the pod changes after restart in Kubernetes, so the rpc call gets stuck because the old ip is unreachable.
To solve this problem, do not cache ReplicationGroupAdmin. The usage of ReplicationGroupAdmin is not frequent, so the overhead of build connection can be ignored.

Signed-off-by: gengjun-git <gengjun@starrocks.com>
(cherry picked from commit d3703724815f9871f69c8f8da2a8a2c6dbc270f2)

# Conflicts:
#	fe/fe-core/src/test/java/com/starrocks/ha/BDBHATest.java