[BugFix] Remove the hiveclient from client pool when occuring error (#17697)
The major change in this pr is remove the hiveclient from client pool when there is an error in accessing the hive metastore.
The connection which accessing hive metastore is a long connection. If anything goes wrong on the server that causes a exception throw, we'll retry and put the connection back in the pool. There are a lot of reasons why accessing hive metastore with this long connection causes an error. we use the user case for example.
The hive.metastore.uris of catalog properties be passed a nginx like waggle-dance hostname. If there is an exception by ip changed of user hive metastore, we will use the all uris in hive.metastore.uris to reconnect service. But at this point, we will only use the old ip resolved when the client was initialized. By default, hive metastore expects all available ip addresses to be passed as arguments hive.metastore.uris. So we resolve uris for each reconnect. For all other exceptions, we added the behavior to remove the hive client from the pool, ensuring that we don't reuse the wrong connection for every accessing.
The only limitation of each resove url is that InetAddress may reuse ip according to the hostname, but the ip cache ttl of jvm defaults to 30 seconds, which can be ignored in corner case.

(cherry picked from commit 92c07fe79635464dbbf5e524225661ab5dd5fbd5)