[BugFix] Fix the bug that the primary key table supports mixed upserts and deletes causing BE crash  (#17793)
This bug was introduced by the pr(https://github.com/StarRocks/starrocks/pull/17264) which add support  for mix upsert and delete for primary key table.

The pr introduced the `_delfile_idxes` to save the delete file idxes but ignore the replicated storage engine.  It only update the `_num_delfile` but not update `_delfile_idxes`. So BE will crash in the following check `DCHECK(_delfile_idxes.size() == _num_delfile);`

Signed-off-by: zhangqiang <qiangzh95@gmail.com>