[BugFix] Fix the bug that BE may crash when separating primary key and sortkey (#17740)
If we separate primary key and sort key when creating a table, we will store data in order of sort key. And when we execute query, we will use sort key to get the data scan range.

So in the following code, we will get seek range of segment and the `input` keep the data in order of sort key, so we should
use sort key column schema to convert `input` to get the right scan range.
```
Status TabletReader::_to_seek_tuple(const TabletSchema& tablet_schema, const OlapTuple& input, SeekTuple* tuple,
                                    MemPool* mempool) {
    ...
    for (size_t i = 0; i < input.size(); i++) {
        int idx = sort_key_idxes.empty() ? i : sort_key_idxes[i];
        auto f = std::make_shared<Field>(ChunkHelper::convert_field(idx, tablet_schema.column(idx)));
        ...
        if (f->type()->type() == TYPE_CHAR) {
            RETURN_IF_ERROR(datum_from_string(get_type_info(TYPE_VARCHAR).get(), &values.back(), input.get_value(i),
                                              mempool));
        } else {
            RETURN_IF_ERROR(datum_from_string(f->type().get(), &values.back(), input.get_value(i), mempool));
        }
    }
    *tuple = SeekTuple(std::move(schema), std::move(values));
    return Status::OK();
}
```

Signed-off-by: zhangqiang <qiangzh95@gmail.com>