[BugFix] Can not find the correct repo when repo prefix is palo after fe restart (#18015)
Problem:
Now we use __starrocks prefix for repo, but if user have the __palo prefix repo, fe will get wrong
prefix as __starrocks when fe restart.

Solution:
When Fe recover meta data for repo, it should try to list __palo prefix repo to check that if  __palo prefix repo
is exist or not and reset prefix to the correct value. In this way, we can avoid to persist the prefix  value into the Image File.

Signed-off-by: srlch <linzichao@starrocks.com>