[BugFix] BE maybe crash when we delete all records in PrimaryKey table with persistent index in branch-2.5+ (#18623)
This bug is introduced by the pr(#15729) which enable flush persistent index in advance.

If we delete a record and we will keep the record which value is NullIndexValue. In following _merge_compaction, we will merge them and skip the old record if new record value is NullIndexValue. If we import a lot of data at once, we will select some tmp l1 file do merge compaction in advance and the delete operation in these files will be skip, but there are maybe some record in older l1 files which should be delete. So the delete operation will lost in merge compaction advance. In the following merge compaction, we will create an empty shard if all records in this shard are delete, but we maybe read some expired record which should be delete which will cause BE crash.