Support iceberg metadata two level cache (#18730)
#18650
Iceberg may have a lot of matadata files, and in this scenario network io may be critical, we add two level cache so that cache metadata on local disk.
All iceberg catalogs share this cache, when memory cache is not full, new data will be loaded and cached in memory, when memory cache is full or data expired, data retreat to disk.

Signed-off-by: blanklin030 <luis01@foxmail.com>