[Enhancement] Default replication_num based on run mode (#18660)
- If run mode is shared_nothing, set default replication_num to 3
- If run mode is shared_data, set default replication_num to 1

For table with decoupled compute and storage, if user specify a value of "replication_num"
greater than 1, will NOT really create multiple replicas of data. However, the check of
"replication_num" on many code paths still takes effect, such as the check of the number of live backends.

For table with decoupled compute and storage, the show create table command will still display
the value specified by the user, although the value has no actual effect.

Signed-off-by: sduzh <zhuming9011@qq.com>