[Enhancement] Reduce S3 headObject requests in lake table's GC
Add a new FileSystem interface iterate_dir2() for getting
the file modification time while traversing the directory
to avoid additional `FileSystem::get_file_modified_time()`
calls.

In fact, this PR will not have any optimization, because the
new interface is not really implemented in the StarletFilesystem,
another PR is needed.

Signed-off-by: sduzh <zhuming9011@qq.com>