[BugFix] Support concrete scalar types as return column types of polymorphic TVFs (#17357)
IF the TVF is declared as returning both polymorphic column types and concrete types, the concrete types right now are dropped when resolving the type.
For example:

FUNC(ANY_ARRAY) -> <ANY_ELEMENT, BIGINT> is the template,

and the invocation is
FUNC(ARRAY_VARCHAR).

During resolution, we correctly resolve ANY_ELEMENT as VARCHAR, but we also want to retain BIGINT as one of the return columns.

---------

Signed-off-by: Andrey Gubichev <a.gubichev@celonis.de>
Signed-off-by: blanklin030 <luis01@foxmail.com>