[Enhancement] Add session variables: max_parallel_scan_instance_num (#19024) (#19233) (#19242)
The parallelism number of scan cannot be adjusted manually. The parallelism number of automatically calculated is sometimes very high, taking up a lot of cpu、memory or io resources, but the performance improvement is not obvious. So we add a session variable to limit the maximum parallelism of scan. It's useful for restrict the resource of large scan query.

Signed-off-by: trueeyu <lxhhust350@qq.com>