[Enhancement] Rectify some behaviors when auto creating tablet (#19062)
1. When the backnum is smaller, try to creating more tablets
```
if (backendNum <= 12) {
    bucketNum = 2 * backendNum;
} else if (backendNum <= 24) {
    bucketNum = (int) (1.5 * backendNum);
} else if (backendNum <= 36) {
    bucketNum = 36;
} else {
    bucketNum = Math.min(backendNum, 48);
}
```

2. Change the behavior when user specify the tablet num
(1) If the user has specified the tablet num, StarRocks will comply the
number, ignore the Config.enable_auto_tablet_distribution
(2) If enable_auto_tablet_distribution = true, AUTO Tablet Distibution
    also take effect on exist table when adding new partitions

3. ColocateTable should comply the original tablet num when creating table.
ColocateTable should keep the tablet num is same among partitions.
So the auto tablet distribution is not suitable for this case.

4. Remove the dynamic_partition.buckets
After the AUTO Tablet Distibution, the dynamic_partition.buckets is not
necessary. We can auto determine the number of tablet