[Enhancement] Simplify case-when to leverage predicates push-down to storage engine (#19407)
* [Enhancement] Simplify case-when to leverage predicates push-down to storage engine

Signed-off-by: satanson <ranpanf@gmail.com>
----
Reduce predicates containing case-when into simple predicates that can be leveraged by storage engine to speedup data access.
Suppor predicates:

in-predicates: case-when-expr in/not in (constants),
eq-predicates: case-when-expr =/<> constant,
is-null-predicates, case-when-expr is/is not null,
Case-when-expr in predicates above must output constant values, some if and nullif expressions are also converted into case-when-expr to adopt this simplification.

Case-when-expr that has case-clause often can be simplified into simple form, while the ones has no case-clause can not, so for no-case case-when-exprs, only when rhs constant values of predicates hit the top 2 branches of the case-when-expr, the simplification would take place.

Case-when that is absent of else-clause or can output NULL values, are apt to buggy, so in some cases, such case-when are not simplified.

Signed-off-by: blanklin030 <luis01@foxmail.com>