BugFix: fix primary key table may have concurrent compaction (#19692)
The commit phase and apply phase of primary key table compaction is asynchronous, and we use _compaction_running to prevent multiple compaction tasks from being executed on the same table.

However, the _compaction_running will be set false after commit finish and it is possible that the tablet is selected to execute compaction again while the last compaction task has not yet finish the apply. The two compaction tasks may be concurrent and modify the _compaction_state at the same time which may cause unexpected error.

This is just a potential bug, not sure if it has caused the problem.

Signed-off-by: zhangqiang <qiangzh95@gmail.com>
(cherry picked from commit 31c7706fedee41ebe1d716fa711c5ef472dddf25)

# Conflicts:
#	be/src/storage/tablet_updates.cpp