[Enhancement] Add primary index cache memory limitation (#18392)
The primary index of PrimaryKey table maybe cost a lot of memory, so we introduce persistent index to limit the memory usage. However, persistent index will keep the index in memory when the usage is very small. So if we have a lot of tablets and all persistent index will be save in the index cache which will cost a lot of memory.

This pr add a soft limit primary_index_memory_limit of primary index memory usage, if the memory usage is overlimit, we will try to find the unused primary index and release them.

Signed-off-by: zhangqiang <qiangzh95@gmail.com>
(cherry picked from commit 3a624f43ade26721af064f321998a5850ac51b45)

# Conflicts:
#	be/src/http/action/update_config_action.cpp
#	be/src/storage/persistent_index.h