[Enhancement] Report error state tablet to FE and show error state tablet in statistic (#19517)
To avoid some serious bugs, we add many check for primary key tablet and if check failed, primary key tablet state will be set error which will refuse all operation. However, the error state is not easily observable and this pr report the error state tablet to FE and we can use show proc '/statistic/; to check if there are some tablet in error state.

Signed-off-by: zhangqiang <qiangzh95@gmail.com>