[BugFix] Be crash when init AUTO_INCREMENT map and Delete op lead to id wasting (#19526) (#19527)
Problem 1:
When a new auto increment map for a table init in BE, BE may crash because of this flag, there will be a race condition, which will lead to wrong id assignment, and eventually crash because the vector is out of bounds.

Problem 2:
When using stream load with partial_update = false and delete flag for some data, the chunk in OlapTableSink::send_chunk contain the data which will be delete. And auto increment id will also allocated for those rows means that delete op will waste the auto increment id in this case.

Solution:
1. Check if the row has op = delete in OlapTableSink::send_chunk
2. The need_init is not needed now, we can remove it safely

Signed-off-by: srlch <linzichao@starrocks.com>
(cherry picked from commit 1fa0b1770b5e79a3bd3e42bbab3d21d2cffa17c7)