[BugFix] auto increment insert crash (#20329)
Problem:
This problem is introduced by commit 12b61a4dc89682189386e612f7d5fcb55be67555. This commit want to check the op column if the loading
contain delete_op. But for the normal INSERT statement, there is no op column and the code here does not check if the col exist or not so led to the crash.

Solution:
check the op column exist or not

Signed-off-by: srlch <linzichao@starrocks.com>