[BugFix] streamload crash for table with auto increment column and without op column (#20431) (#20435)
Problem:
This problem is introduced by commit 12b61a4dc896821. The op column's type is different in specified case and not specified case. So the nullable column type cast may be fail when we do not specify column.

Solution:
raw_data() can be called by the base type class Column and we do not cast the derive type.

Signed-off-by: srlch <linzichao@starrocks.com>