[BugFix] Remove be shared scan configs
Currently, there are two types of shared scan configs: fe session
variable and be configs. the default value of session variable is false,
and be config for connector scan node is true, this will cause data
skew, such as scan operator uses SharedMorselQueueFactory and kDirect
chunk buffer.

This pr fixes:
1. Only use session variable enable_shared_scan to control shared scan.
2. Create morsel queue factory according to enable_shared_scan variable.
3. Make cloud native table behavior same as local table.

Signed-off-by: wyb <wybb86@gmail.com>