[BugFix] some bug fix for auto increment (#20575) (#20576)
Problem 1:
Currently, auto increment column does not support modify a column into a AUTO_INCREMENT column. But it is possible that modify a AUTO_INCREMENT column into a normal column. This should be abort.

Problem 2:
There is no mechanism to handle the overflow of the auto increment id counter, which will cause be crash.

Problem 3:
The planning for insert into is not feasible when the auto increment is not specified and defined before other normal column.

Problem 4:
When we recover a table, the auto increment id will restart from 0, which is a problem.

Solution:
1. Abort the case that we modify a AUTO_INCREMENT column into a normal column.
2. Throw an exception when the id counter is overflow.
3. Fix the planning problem for insert statement.
4. remove the auto increment id when truly erase the table.
5. add some test case for auto increment column.

Signed-off-by: srlch <linzichao@starrocks.com>
(cherry picked from commit da6f58aba530ac4a68f5c2681e5a8a8332187339)