[BugFix]Fix decommission backend bug (#20165)
Introduced by #19635
handleReplicaRelocating will try to fix the version incomplete replica first, if no version incomplete replica, it will go to the handleReplicaMissing logic. This PR #19635 changed the exception to UNRECOVERABLE if there is no version incomplete replica, it's ok for the process of VERSION_INCOMPLETE, but for the REPLICA_RELOCATING, it won't go to the handleReplicaMissing logic, because this exception is not caught.
Catch the UNRECOVERABLE exception when process the replica relocation schedule.