[BugFix] Fix unstable TabletBinlogTest (#20767)
`TabletBinlogTest#test_generate_binlog` is unstable, and will fail because of the following exception. It's the problem of the test, not the problem of binlog logic code.

```
[ RUN      ] TabletBinlogTest.test_generate_binlog
/root/starrocks/be/test/storage/binlog_test_base.cpp:139: Failure
Expected equality of these values:
  expect->timestamp
    Which is: 1680230581000000
  actual->timestamp_in_us
    Which is: 1680230582000000
/root/starrocks/be/test/storage/binlog_test_base.cpp:139: Failure
Expected equality of these values:
  expect->timestamp
    Which is: 1680230581000000
  actual->timestamp_in_us
    Which is: 1680230582000000
/root/starrocks/be/test/storage/binlog_test_base.cpp:139: Failure
Expected equality of these values:
  expect->timestamp
    Which is: 1680230581000000
  actual->timestamp_in_us
    Which is: 1680230582000000
```

The exception happens when verifying the timestamp of binlog is as expected. We get the timestamp of binlog from `Rowset#creation_time()`. The return values could be different before and after making rowset visible in  `Tablet#add_inc_rowset`, and we actually need the one after it's visible. But the test calls `Rowset#creation_time()` before the rowset is visible, and think it's the expected timestamp which leads to the test failure. 
 Because the time unit of `Rowset#creation_time()`  is second, so it's possible that the timestamps could be same before and after the rowset is visible if `add_inc_rowset` executes fast. And this is why the test is unstable.

Signed-off-by: PengFei Li <lpengfei2016@gmail.com>