[Enhancement] Improve disk balance speed when added multi backends (#19418)
When doing cluster disk balancing, most of the tablets selected are from the same BE, because of the limitation of IO and network, the migration is very slow.
To improve the migration parallelism, we divide the BE into two groups: one is higher than the average disk usage, and another is lower than the average disk usage. When selecting the tablet, we use round robin to change the source BE and target BE once a tablet is chosen.

Signed-off-by: gengjun-git <gengjun@starrocks.com>
(cherry picked from commit 81c69fd5ac220ae5ab5a614048d6118bbefc653c)

# Conflicts:
#	fe/fe-core/src/main/java/com/starrocks/common/Config.java