[Enhancement] Improve refresh external table performance (#21265)
Signed-off-by: stephen <stephen5217@163.com>

Currently we get the partitions that need to be refreshed by filtering the partitions of all tables in the cache when refreshing. One advantage of this is that partitions cache that do not exist in the table of user side can be cleaned up for each refresh manually. This can solve the issue of inconsistency in the partition cache of the table with the same name rebuilt on the hive side.
By @DorianZheng advice, we will provide background auto refresh feature in follow version. this will cause one cpu of fe node to be full frequently. So we only refresh partition existed in hms. But this won't clean up the partition cache which doesn't exist in hms. it’s a trade-off.

(cherry picked from commit 05e8f5ec90264218de65dee527bc374de642e987)