[BugFix] BE stream load rollback failure (#21127)
When the BE stream load executor initiates a txn rollback, the BE doesn't include table info. The FE checks the authorization with user table info and database info. When the user's LOAD grant is granted with the specific table, the authorization check will fail, and Access denied; you need (at least one of) the LOAD privilege(s) for this operation would be returned.