[Enhancement] adjust the transaction publish thread pool size strategy (#21368)
In the current implementation, the transaction publish thread pool adopts a fixed number of threads strategy, with the number of threads being set to the configuration item transaction_publish_version_worker_count=8. However, this may cause problems because the default 8 threads may not be sufficient when the import concurrency is high, and can only be adjusted manually by analyzing and adjusting the number of threads.

Therefore, here comes a more friendly strategy, which is to take advantage of the dynamic thread pool by setting the minimum number of transaction publish threads to 8, and the maximum number of threads to half of the machine's cpu count by default. At the same time, it allows dynamic adjustment of the thread pool's maximum threads through an HTTP interface.

(cherry picked from commit 6691fc0636b58c525f4d6201b0ca1ca78ec5e042)

# Conflicts:
#	be/src/agent/agent_server.cpp